DetailViewController decomposition
