+++ SimbirSoft/.swiftlint.yml
@@ -1,6 +1,7 @@
 disabled_rules:
 - trailing_whitespace
 - line_length
+- type_body_length
 
 excluded:
 - Pods
