+++ SimbirSoft/SimbirSoft/HelpModule/DetailScreen/MainView/Views/InfoViews/PhoneView.swift
@@ -1,0 +1,76 @@
+//
+//  PhoneView.swift
+//  SimbirSoft
+//
+//  Created by NaYfront on 15.08.2022.
+//
+
+import UIKit
+import SnapKit
+
+class PhoneView: UIView {
+    // MARK: - Private Properties
+    private let event: Event
+    
+    // MARK: - User Interface
+    lazy var phoneImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.contentMode = .scaleAspectFill
+        imageView.clipsToBounds = true
+        imageView.image = #imageLiteral(resourceName: "iconPhone")
+        
+        return imageView
+    }()
+    
+    lazy var phoneLabel: UILabel = {
+        let label = UILabel()
+        label.numberOfLines = 0
+
+        label.attributedText = .toAttributedString(
+            attributes: [UIFont.sfuitextRegular(size: 15),
+                         UIColor.charcoalGrey,
+                         NSMutableParagraphStyle(alignment: .left, minimumLineLength: nil)],
+            text: event.phoneNumber)
+
+        return label
+    }()
+    
+    // MARK: - Initializers
+    init(event: Event) {
+        self.event = event
+        super.init(frame: .zero)
+        
+        setupUI()
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    // MARK: - Private Functions
+    
+    private func addSubviews() {
+        self.addSubview(phoneImageView)
+        self.addSubview(phoneLabel)
+    }
+    
+    private func setupUI() {
+        addSubviews()
+        
+        phoneImageView.snp.makeConstraints { make in
+            make.left.equalToSuperview()
+            make.top.equalToSuperview().offset(3)
+            make.height.width.equalTo(16)
+        }
+
+        phoneLabel.snp.makeConstraints { make in
+            make.right.top.bottom.equalToSuperview()
+            make.left.equalTo(phoneImageView.snp.right).offset(10)
+        }
+
+        self.snp.makeConstraints { make in
+            make.height.equalTo(phoneLabel)
+            make.width.equalTo(phoneLabel).offset(26)
+        }
+    }
+}
